refactor logic into reusable function and add dev test
